+++ rust/cubesql/cubesql/src/sql/database_variables/postgres/session_vars.rs
@@ -1,109 +1,77 @@
 use datafusion::scalar::ScalarValue;
-use std::collections::HashMap;
 
 use crate::compile::{DatabaseVariable, DatabaseVariables};
 
-pub fn defaults() -> DatabaseVariables {
-    let mut variables: DatabaseVariables = HashMap::new();
+pub const CUBESQL_PENALIZE_POST_PROCESSING_VAR: &str = "cubesql_penalize_post_processing";
 
-    variables.insert(
-        "client_min_messages".to_string(),
+pub fn defaults() -> DatabaseVariables {
+    let variables = [
         DatabaseVariable::system(
             "client_min_messages".to_string(),
             ScalarValue::Utf8(Some("NOTICE".to_string())),
             None,
         ),
-    );
-
-    variables.insert(
-        "timezone".to_string(),
         DatabaseVariable::system(
             "timezone".to_string(),
             ScalarValue::Utf8(Some("GMT".to_string())),
             None,
         ),
-    );
-
-    variables.insert(
-        "application_name".to_string(),
         DatabaseVariable::system(
             "application_name".to_string(),
             ScalarValue::Utf8(None),
             None,
         ),
-    );
-
-    variables.insert(
-        "extra_float_digits".to_string(),
         DatabaseVariable::system(
             "extra_float_digits".to_string(),
             ScalarValue::UInt32(Some(1)),
             None,
         ),
-    );
-
-    variables.insert(
-        "transaction_isolation".to_string(),
         DatabaseVariable::system(
             "transaction_isolation".to_string(),
             ScalarValue::Utf8(Some("read committed".to_string())),
             None,
         ),
-    );
-
-    variables.insert(
-        "max_allowed_packet".to_string(),
         DatabaseVariable::system(
             "max_allowed_packet".to_string(),
             ScalarValue::UInt32(Some(67108864)),
             None,
         ),
-    );
-
-    variables.insert(
-        "max_index_keys".to_string(),
         DatabaseVariable::system(
             "max_index_keys".to_string(),
             ScalarValue::UInt32(Some(32)),
             None,
         ),
-    );
-
-    variables.insert(
-        "lc_collate".to_string(),
         DatabaseVariable::system(
             "lc_collate".to_string(),
             ScalarValue::Utf8(Some("en_US.utf8".to_string())),
             None,
         ),
-    );
-
-    variables.insert(
-        "standard_conforming_strings".to_string(),
         DatabaseVariable::system(
             "standard_conforming_strings".to_string(),
             ScalarValue::Utf8(Some("on".to_string())),
             None,
         ),
-    );
-
-    variables.insert(
-        "max_identifier_length".to_string(),
         DatabaseVariable::system(
             "max_identifier_length".to_string(),
             ScalarValue::UInt32(Some(63)),
             None,
         ),
-    );
-
-    variables.insert(
-        "role".to_string(),
         DatabaseVariable::system(
             "role".to_string(),
             ScalarValue::Utf8(Some("none".to_string())),
             None,
         ),
-    );
+        // Custom cubesql variables
+        DatabaseVariable::user_defined(
+            CUBESQL_PENALIZE_POST_PROCESSING_VAR.to_string(),
+            ScalarValue::Boolean(Some(false)),
+            None,
+        ),
+    ];
+
+    let variables = IntoIterator::into_iter(variables)
+        .map(|v| (v.name.clone(), v))
+        .collect::<DatabaseVariables>();
 
     variables
 }